Moves test syscalls and test instruction meter to common.rs
